+++ demo/app/javascript/packs/application.js
@@ -1,1 +1,4 @@
+import Rails from "@rails/ujs"
+Rails.start()
+
 import "controllers"
