+++ demo/package.json
@@ -2,6 +2,7 @@
   "name": "demo",
   "private": true,
   "dependencies": {
+    "@rails/ujs": "^6.1.4",
     "@rails/webpacker": "5.4.0",
     "polaris-view-components": "^0.1.0",
     "stimulus": "^2.0.0",
