+++ demo/test/components/previews/rails/helpers_component_preview.rb
@@ -1,0 +1,4 @@
+class Rails::HelpersComponentPreview < ViewComponent::Preview
+  def button_to
+  end
+end
