+++ lib/polaris/view_components/engine.rb
@@ -20,6 +20,7 @@ module Polaris
       initializer "polaris_view_components.helpers" do
         ActiveSupport.on_load(:action_controller_base) do
           helper Polaris::ViewHelper
+          helper Polaris::UrlHelper
           helper Polaris::ActionHelper
           helper Polaris::ConditionalHelper
         end
